рефакторинг bigUntestableCode
