+++ package.json
@@ -20,7 +20,7 @@
     "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "14.2.1",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^25.2.2",
+    "eslint-plugin-jest": "^25.2.4",
     "jest": "^27.3.1",
     "lint-staged": "^11.1.2",
     "sass": "^1.41.1",
