+++ src/css/style.scss
@@ -1,0 +1,118 @@
+#app {
+    font-family: Avenir, Helvetica, Arial, sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    text-align: center;
+    color: #2c3e50;
+    margin-top: 60px;
+}
+
+.app{
+    width: 100%;
+    height: 100vh;
+    // display: none;
+    &__btn {
+        width: 100px;
+        height: 50px;
+        font-size: 15px;
+        border: 0;
+        border-radius: 10px;
+        margin: 0 15px;
+        font-weight: 600;
+        cursor: pointer;
+    }
+
+    &__btn--yellow {
+        background: rgb(255, 246, 121);
+    }
+
+    &__btn--green {
+        background: rgb(184, 255, 184);
+    }
+
+    &__btn--blue {
+        background: rgb(151, 151, 255);
+    }
+
+    &__error {
+        background: rgb(255, 113, 113);
+        font-weight: 600;
+        font-size: 22px;
+        height: 100px;
+        border-radius: 10px;
+        text-align: center;
+    }
+
+    &__loader-wrapper {
+        background: #2c3e50;
+        width: inherit;
+        height: inherit;
+        position: absolute;
+        top: 0;
+        left: 0;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        
+        .loader {
+            display: inline-block;
+            width: 30px;
+            height: 30px;
+            position: relative;
+            border: 4px solid #Fff;
+            animation: loader 2s infinite ease;
+          }
+          
+          .loader-inner {
+            vertical-align: top;
+            display: inline-block;
+            width: 100%;
+            background-color: #fff;
+            animation: loader-inner 2s infinite ease-in;
+          }
+          
+          @keyframes loader {
+            0% {
+              transform: rotate(0deg);
+            }
+            
+            25% {
+              transform: rotate(180deg);
+            }
+            
+            50% {
+              transform: rotate(180deg);
+            }
+            
+            75% {
+              transform: rotate(360deg);
+            }
+            
+            100% {
+              transform: rotate(360deg);
+            }
+          }
+          
+          @keyframes loader-inner {
+            0% {
+              height: 0%;
+            }
+            
+            25% {
+              height: 0%;
+            }
+            
+            50% {
+              height: 100%;
+            }
+            
+            75% {
+              height: 100%;
+            }
+            
+            100% {
+              height: 0%;
+            }
+          }
+    }
+}
