+++ src/js/loader.js
@@ -1,0 +1,5 @@
+const loader = document.querySelector('#loader');
+
+export const loaderDisplayBlock = () => loader.style.display = 'block';
+
+export const loaderDisplayNone = () => loader.style.display = 'none';
